Refactor: tokenAccess, userEmail, userImage => backState로 옮기기 	- 위의 state 들은 api를 통해 받아오는 것이기 때문
